requirements.txt
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
bcrypt==4.2.1
billiard==4.2.1
boto3==1.35.68
botocore==1.35.68
celery==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.5.3
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.17
pytz==2025.1
PyYAML==6.0.2
redis==6.0.0
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
